avoid duplicate ImageStream
The S2I templates for `greeting-service` and `name-service` used to
define an `ImageStream` which was in both cases called `runtime`.
When using the S2I templates manually for deploying and later
deleting the OpenShift resources, this caused an unnecessary error
when deleting the `runtime` image stream for the 2nd time.

This commit provides a simple solution: it renames the `ImageStream`
in the `name-service` from `runtime` to `runtime-name`.